fix some weird stuff

+++ nix/mkNeovim.nix
@@ -1,18 +1,19 @@
 # Function for creating a Neovim derivation
 {
-  pkgs,
   lib,
   stdenv,
+  sqlite,
+  git,
+  neovim-unwrapped,
   # Set by the overlay to ensure we use a compatible version of `wrapNeovimUnstable`
-  pkgs-wrapNeovim ? pkgs,
+  wrapNeovimUnstable,
+  neovimUtils,
 }:
 with lib;
 {
   # NVIM_APPNAME - Defaults to 'nvim' if not set.
   # If set to something else, this will also rename the binary.
   appName ? null,
-  # The Neovim package to wrap
-  neovim-unwrapped ? pkgs-wrapNeovim.neovim-unwrapped,
   plugins ? [ ], # List of plugins
   # List of dev plugins (will be bootstrapped) - useful for plugin developers
   # { name = <plugin-name>; url = <git-url>; }
@@ -32,8 +33,11 @@ with lib;
   withSqlite ? true, # Add sqlite? This is a dependency for some plugins
   # You probably don't want to create vi or vim aliases
   # if the appName is something different than "nvim"
-  viAlias ? appName == "nvim", # Add a "vi" binary to the build output as an alias?
-  vimAlias ? appName == "nvim", # Add a "vim" binary to the build output as an alias?
+  # Add a "vi" binary to the build output as an alias?
+  viAlias ? appName == null || appName == "nvim",
+  # Add a "vim" binary to the build output as an alias?
+  vimAlias ? appName == null || appName == "nvim",
+  wrapRc ? true,
 }:
 let
   # This is the structure of a plugin definition.
@@ -48,14 +52,14 @@ let
     runtime = { };
   };
 
-  externalPackages = extraPackages ++ (optionals withSqlite [ pkgs.sqlite ]);
+  externalPackages = extraPackages ++ (optionals withSqlite [ sqlite ]);
 
   # Map all plugins to an attrset { plugin = <plugin>; config = <config>; optional = <tf>; ... }
   normalizedPlugins = map (x: defaultPlugin // (if x ? plugin then x else { plugin = x; })) plugins;
 
   # This nixpkgs util function creates an attrset
   # that pkgs.wrapNeovimUnstable uses to configure the Neovim build.
-  neovimConfig = pkgs-wrapNeovim.neovimUtils.makeNeovimConfig {
+  neovimConfig = neovimUtils.makeNeovimConfig {
     inherit
       extraPython3Packages
       withPython3
@@ -101,11 +105,17 @@ let
     '';
 
     installPhase = ''
-      cp -r after $out/after
-      rm -r after
       cp -r lua $out/lua
       rm -r lua
-      cp -r * $out/nvim
+      # Copy nvim/after only if it exists
+      if [ -d "after" ]; then
+          cp -r after $out/after
+          rm -r after
+      fi
+      # Copy rest of nvim/ subdirectories only if they exist
+      if [ ! -z "$(ls -A)" ]; then
+          cp -r -- * $out/nvim
+      fi
     '';
   };
 
@@ -115,7 +125,6 @@ let
   # It also adds logic for bootstrapping dev plugins (for plugin developers)
   initLua =
     ''
-      vim.loader.enable()
       -- prepend lua directory
       vim.opt.rtp:prepend('${nvimRtp}/lua')
     ''
@@ -132,7 +141,7 @@ let
         dev_plugin_path = dev_plugins_dir .. '/${plugin.name}'
         if vim.fn.empty(vim.fn.glob(dev_plugin_path)) > 0 then
           vim.notify('Bootstrapping dev plugin ${plugin.name} ...', vim.log.levels.INFO)
-          vim.cmd('!${pkgs.git}/bin/git clone ${plugin.url} ' .. dev_plugin_path)
+          vim.cmd('!${git}/bin/git clone ${plugin.url} ' .. dev_plugin_path)
         end
         vim.cmd('packadd! ${plugin.name}')
       '') devPlugins
@@ -156,9 +165,9 @@ let
     # Add external packages to the PATH
     ++ (optional (externalPackages != [ ]) ''--prefix PATH : "${makeBinPath externalPackages}"'')
     # Set the LIBSQLITE_CLIB_PATH if sqlite is enabled
-    ++ (optional withSqlite ''--set LIBSQLITE_CLIB_PATH "${pkgs.sqlite.out}/lib/libsqlite3.so"'')
+    ++ (optional withSqlite ''--set LIBSQLITE_CLIB_PATH "${sqlite.out}/lib/libsqlite3.so"'')
     # Set the LIBSQLITE environment variable if sqlite is enabled
-    ++ (optional withSqlite ''--set LIBSQLITE "${pkgs.sqlite.out}/lib/libsqlite3.so"'')
+    ++ (optional withSqlite ''--set LIBSQLITE "${sqlite.out}/lib/libsqlite3.so"'')
   );
 
   luaPackages = neovim-unwrapped.lua.pkgs;
@@ -179,7 +188,7 @@ let
       }"'';
 
   # wrapNeovimUnstable is the nixpkgs utility function for building a Neovim derivation.
-  neovim-wrapped = pkgs-wrapNeovim.wrapNeovimUnstable neovim-unwrapped (
+  neovim-wrapped = wrapNeovimUnstable neovim-unwrapped (
     neovimConfig
     // {
       luaRcContent = initLua;
@@ -191,7 +200,7 @@ let
         + extraMakeWrapperLuaCArgs
         + " "
         + extraMakeWrapperLuaArgs;
-      wrapRc = true;
+      wrapRc = wrapRc;
     }
   );
 
@@ -204,4 +213,5 @@ neovim-wrapped.overrideAttrs (oa: {
     + lib.optionalString isCustomAppName ''
       mv $out/bin/nvim $out/bin/${lib.escapeShellArg appName}
     '';
+  meta.mainProgram = if isCustomAppName then appName else oa.meta.mainProgram;
 })
